Refactor tests to modules
